requirements.txt
boto3
awscli
pypdf
langchain
streamlit
faiss-cpu
langchain_community
pdfplumber